+++ .vscode/settings.json
@@ -6,6 +6,7 @@
         "**/.hg": true,
         "**/.DS_Store": true,
         "**/*.js": { "when": "$(basename).ts"},
-        "**/*.js.map": { "when": "$(basename)"}
+        "**/*.js.map": { "when": "$(basename)"},
+        "**/*.d.ts": { "when": "$(basename).ts"}
     }
 }
